Move all subscribers to their own module
