+++ awscli/customizations/s3/s3handler.py
@@ -35,15 +35,13 @@ from awscli.customizations.s3.results import ResultProcessor
 from awscli.customizations.s3.results import CommandResultRecorder
 from awscli.customizations.s3.utils import RequestParamsMapper
 from awscli.customizations.s3.utils import StdoutBytesWriter
-from awscli.customizations.s3.utils import ProvideSizeSubscriber
-from awscli.customizations.s3.utils import ProvideUploadContentTypeSubscriber
-from awscli.customizations.s3.utils import ProvideCopyContentTypeSubscriber
-from awscli.customizations.s3.utils import ProvideLastModifiedTimeSubscriber
-from awscli.customizations.s3.utils import CopyPropsSubscriberFactory
-from awscli.customizations.s3.utils import DirectoryCreatorSubscriber
-from awscli.customizations.s3.utils import DeleteSourceFileSubscriber
-from awscli.customizations.s3.utils import DeleteSourceObjectSubscriber
-from awscli.customizations.s3.utils import DeleteCopySourceObjectSubscriber
+from awscli.customizations.s3.subscribers import (
+    ProvideSizeSubscriber, ProvideUploadContentTypeSubscriber,
+    ProvideCopyContentTypeSubscriber, ProvideLastModifiedTimeSubscriber,
+    CopyPropsSubscriberFactory, DirectoryCreatorSubscriber,
+    DeleteSourceFileSubscriber, DeleteSourceObjectSubscriber,
+    DeleteCopySourceObjectSubscriber
+)
 from awscli.compat import get_binary_stdin
